// packages/api/src/gateway/models/GuidPaginationResultDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type GuidPaginationResultDto = {
    currentPage?: number;
    pageSize?: number;
    totalPages?: number;
    totalItems?: number;
    data?: Array<string> | null;
};